add task delete task just for admin or for user of taks

+++ src/AppBundle/Controller/TaskController.php
@@ -83,11 +83,12 @@ class TaskController extends AbstractController
      */
     public function deleteTaskAction(Task $task)
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
-
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+        if( in_array("ROLE_ADMIN",$this->getUser()->getRoles()) || $this->getUser()->getUsername() === $task->getCreator()->getUsername() ){
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($task);
+            $em->flush();
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+        }
 
         return $this->redirectToRoute('task_list');
     }
